refactor pycnocline work as COAsT class
