+++ coast/__init__.py
@@ -9,4 +9,5 @@ from .CDF import CDF
 from .CRPS import CRPS
 from .interpolate_along_dimension import interpolate_along_dimension
 from .STATS import STATS
+from .DIAGNOSTICS import DIAGNOSTICS
 #from .ERRORS import ERRORS
